convert stop bits to float

+++ image/src/app/module/main.py
@@ -28,7 +28,7 @@ def module_main():
         str_data_bytes = bytes(str(APPLICATION['DATA_BITS']), "ascii")
         data_bytes= int(str_data_bytes)
         str_stop_bits=bytes(str(APPLICATION['STOP_BITS']), "ascii")
-        stop_bits=int(str_stop_bits)
+        stop_bits=float(str_stop_bits)
         ser = serial.Serial(APPLICATION['PORT'], APPLICATION['BAUD_RATE'], timeout=1,bytesize=data_bytes,parity=parity_dict.get(APPLICATION['PARITY']), stopbits=stop_bits)
         while True:
             #  read json payload
